implemented editing of existing menus
